moved state to home.js
