+++ src/SearchResults.js
@@ -3,7 +3,10 @@ import { Container, Box, Button } from 'bloomer'
 
 export default class SearchResults extends Component {
 
-
+    test = (e) => {
+        e.preventDefault()
+        console.log(this.props.recipes)
+    }
 
     render() {
         return (
@@ -13,6 +16,7 @@ export default class SearchResults extends Component {
                 </Button>
                 <Container>
                     <Box id="results-box">
+                        <Button onClick={this.test}>test</Button>
                     </Box>
                 </Container>
             </React.Fragment>
